Move the javascript and user-agent helpers into utils.
